rename required attrs fixture app
