+++ test/fixtures/index.js
@@ -6,7 +6,7 @@ import {DelayedAttrRemoveApp} from './delayed-attr-remove-app';
 import {NestedApp, NestedChild} from './nested-app';
 import {NestedKeyedChildrenApp, NestedKeyedChild1, NestedKeyedChild2} from './nested-keyed-children-app';
 import {NestedPartialStateParent, NestedPartialStateChild} from './nested-partial-state-app';
-import {BadBooleanRequiredAttrsSchemaApp, RequiredAttrsSchemaApp} from './required-attrs-schema-app';
+import {BadBooleanRequiredAttrsSchemaApp, RequiredAttrsSchemaApp} from './required-attrs-schema-apps';
 import {ShadowDomApp} from './shadow-dom-app';
 import {SimpleApp} from './simple-app';
