+++ test/fixtures/required-attrs-schema-apps.js
